Rename extension file
